JsonFlatFileDataStoreのサンドボックス環境を分離
